Add files via upload
This perl code is for "Dictionary Word Checker".Dictionary Word Checker can be used to enter program language source statements or to create documents such as technical manuals. In this checker we have appended the dictionary so that the meaning of the word we have entered in the text area is shown in the module.